+++ src/panoramic/cli/errors.py
@@ -346,15 +346,24 @@ class OrphanFieldFileError(ValidationError):
 class MissingFieldFileError(ValidationError):
     field_slug: str
     dataset_slug: str
+    data_source: str
+    data_reference: str
+    identifier: bool
 
     def __init__(
         self,
         *,
         field_slug: str,
         dataset_slug: str,
+        data_source: str,
+        data_reference: str,
+        identifier: bool,
     ) -> None:
         self.field_slug = field_slug
         self.dataset_slug = dataset_slug
+        self.data_source = data_source
+        self.data_reference = data_reference
+        self.identifier = identifier
         super().__init__(f'Missing field file for slug {field_slug} under dataset {dataset_slug}')
 
     def __eq__(self, o: object) -> bool: